Rename context and provider
